# Conflicts:
#	Datapack
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/Config.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/data/database/data/PlayerVariableData.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/instancemanager/InstanceManager.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/model/Clan.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/model/actor/Attackable.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/model/actor/instance/Player.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/model/actor/instance/PlayerFactory.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/ExIncomingPackets.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/clientpackets/UseItem.java
#	Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/settings/GeneralSettings.java
#	Gameserver/src/main/resources/sql/install/player_variables.sql
#	Script/src/main/org.l2j.scripts/org.l2j.scripts/handlers/admincommandhandlers/AdminVitality.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/admincommandhandlers/AdminSayhaGrace.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/admincommandhandlers/AdminVitality.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/effecthandlers/Elixir.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/effecthandlers/RecoverSayhaGraceInPeaceZone.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/effecthandlers/SayhaGraceExpRate.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/effecthandlers/SayhaGracePointUp.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/effecthandlers/SayhaGracePointsRate.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/effecthandlers/SayhaGraceSupport.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/itemhandlers/BlessedScroll.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/itemhandlers/LimitedSayha.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/skillconditionhandlers/CanUseSayhaGraceConsumeItemSkillCondition.java
#	Script/src/main/org.l2j.scripts/org/l2j/scripts/handlers/skillconditionhandlers/ElixirSkillCondition.java

+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/model/holders/ItemChanceHolder.java
@@ -25,14 +25,14 @@ package org.l2j.gameserver.model.holders;
  * @author JoeAlisson
  */
 public class ItemChanceHolder extends ItemHolder {
-    private final double chance;
+    private final float chance;
 
-    public ItemChanceHolder(int id, double chance, long count) {
+    public ItemChanceHolder(int id, float chance, long count) {
         super(id, count);
         this.chance = chance;
     }
 
-    public double getChance() {
+    public float getChance() {
         return chance;
     }
